Style updates for loaders
